server/requirements.txt
flask
paho-mqtt
pyyaml